docs/domain/src/main/java/SpotExchange.java

public class SpotExchange {

    private boolean success;
    private int min;
    private boolean validity;

    public SpotExchange() {
        this.min = 0;
    }

    public boolean isSuccess() {
        return success;
    }

    public void setSuccess(boolean success) {
        this.success = success;
    }

    public int getMin() {
        return min;
    }

    public void setMin(int min) {
        this.min = min;
    }

    public boolean isValidity() {
        return validity;
    }

    public void setValidity(boolean validity) {
        this.validity = validity;
    }

}